# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Apollo Service",
    "description": "Service for searching and enriching leads via the Apollo API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3004"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SearchFilters": {
        "type": "object",
        "properties": {
          "personTitles": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by job titles.",
            "example": [
              "CEO",
              "CTO",
              "VP Engineering"
            ]
          },
          "qOrganizationKeywordTags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by organization keyword tags.",
            "example": [
              "SaaS",
              "fintech"
            ]
          },
          "organizationLocations": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by organization HQ location.",
            "example": [
              "United States",
              "California, US"
            ]
          },
          "organizationNumEmployeesRanges": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "1,10",
                "11,20",
                "21,50",
                "51,100",
                "101,200",
                "201,500",
                "501,1000",
                "1001,2000",
                "2001,5000",
                "5001,10000",
                "10001,"
              ]
            },
            "description": "Filter by employee count ranges.",
            "example": [
              "11,20",
              "21,50",
              "51,100"
            ]
          },
          "qOrganizationIndustryTagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by industry names (use GET /reference/industries for valid values).",
            "example": [
              "Information Technology and Services",
              "Computer Software"
            ]
          },
          "qKeywords": {
            "type": "string",
            "description": "Free-text keyword search across person and organization fields.",
            "example": "machine learning"
          },
          "personLocations": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by person's location (city, state, country). Different from organizationLocations which filters by company HQ.",
            "example": [
              "San Francisco, California, US",
              "New York, US"
            ]
          },
          "personSeniorities": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "entry",
                "senior",
                "manager",
                "director",
                "vp",
                "c_suite",
                "owner",
                "founder",
                "partner"
              ]
            },
            "description": "Filter by seniority level. Valid values: entry, senior, manager, director, vp, c_suite, owner, founder, partner.",
            "example": [
              "director",
              "vp",
              "c_suite"
            ]
          },
          "contactEmailStatus": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "verified",
                "guessed",
                "unavailable",
                "bounced",
                "pending_manual_fulfillment"
              ]
            },
            "description": "Filter by email verification status. Valid values: verified, guessed, unavailable, bounced, pending_manual_fulfillment.",
            "example": [
              "verified"
            ]
          },
          "qOrganizationDomains": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by specific company domains.",
            "example": [
              "google.com",
              "meta.com"
            ]
          },
          "currentlyUsingAnyOfTechnologyUids": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by technology stack (Apollo technology UIDs).",
            "example": [
              "salesforce",
              "hubspot"
            ]
          },
          "revenueRange": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by annual revenue ranges (comma-separated min,max format).",
            "example": [
              "1000000,10000000",
              "10000000,50000000"
            ]
          },
          "organizationIds": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Filter by specific Apollo organization IDs.",
            "example": [
              "5f5e100a01d6b1000169c754"
            ]
          }
        },
        "description": "Search filters. On first call, provide filters to create a cursor at page 1. On subsequent calls, omit to continue from the last page. If provided and different from the stored filters, the cursor resets to page 1."
      },
...
      "SearchParamsResponse": {
        "type": "object",
        "properties": {
          "searchParams": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchFilters"
              },
              {
                "description": "Validated Apollo search filters that returned results."
              }
            ]
          },
          "totalResults": {
            "type": "number",
            "description": "Apollo total_entries for the final search params."
          },
          "attempts": {
            "type": "number",
            "description": "Number of LLM iterations (1 = first try worked)."
          },
          "attemptHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchParamsAttempt"
            },
            "description": "Full history of all attempts for debugging."
          },
          "cached": {
            "type": "boolean",
            "description": "True if the result was served from the 24h cache (no LLM or Apollo calls made)."
          }
        },
        "required": [
          "searchParams",
          "totalResults",
          "attempts",
          "attemptHistory",
          "cached"
        ]
      },
      "SearchParamsAttempt": {
        "type": "object",
        "properties": {
          "searchParams": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchFilters"
              },
              {
                "description": "Apollo search filters. All filters are combined using AND. Start broad and narrow down to avoid empty results."
              }
            ]
          },
          "totalResults": {
            "type": "number"
          }
        },
        "required": [
          "searchParams",
          "totalResults"
        ]
      },
      "SearchParamsRequest": {
        "type": "object",
        "properties": {
          "context": {
            "type": "string",
            "minLength": 1,
            "description": "Unstructured context about the company — website content, target audience description, ICP notes, or any combination. The LLM extracts what it needs.",
            "example": "We are a B2B SaaS company selling developer tools. Target audience: engineering leaders at mid-size tech companies in the US."
          }
        },
        "required": [
          "context"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/search/params": {
      "post": {
        "summary": "Generate Apollo search parameters from context using LLM",
        "description": "Takes unstructured context (website content, target audience, etc.) and uses Claude to generate Apollo search filters. Validates against Apollo — if 0 results, retries with broadened filters (max 10 attempts). Returns validated search params guaranteed to produce results, or the best-effort params after 10 attempts.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Caller's run ID — used as parentRunId when creating child runs",
              "example": "run-abc-123"
            },
            "required": true,
            "description": "Caller's run ID — used as parentRunId when creating child runs",
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID",
              "example": "brand-1"
            },
            "required": true,
            "description": "Brand ID",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID",
              "example": "campaign-1"
            },
            "required": true,
            "description": "Campaign ID",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking",
              "example": "lead-gen"
            },
            "required": false,
            "description": "Feature slug for tracking",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug for run tracking",
              "example": "fetch-lead"
            },
            "required": false,
            "description": "Workflow slug for run tracking",
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchParamsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated and validated search parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchParamsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}